# Prevent crash when cgroups_ref is null in streamEntryIsReferenced() after reload