@kaplanben 2025-08-14
![Logo](https://cdn.ast.checkmarx.net/integrations/logo/Checkmarx.png)
**Checkmarx One – Scan Summary & Details** – [3d285292\-0d25\-4ea4\-a210\-45838580479b](https://gov-il.ast.checkmarx.net/projects/6e34a6db-7b74-4c41-8bc6-d52b8f2a3534/scans?id=3d285292-0d25-4ea4-a210-45838580479b&branch=sundb%2Freload_cgroup_ref)
<!--new issues section--><details><summary><b>New Issues (7)</b></summary>
Checkmarx found the following issues in this Pull Request

Severity|Issue|Source File / Package|Checkmarx Insight
---|---|---|---
![CRITICAL](https://cdn.ast.checkmarx.net/integrations/severity/Critical.png)|[Buffer_Overflow_Wrong_Buffer_Size](https://gov-il.ast.checkmarx.net/results/6e34a6db-7b74-4c41-8bc6-d52b8f2a3534/3d285292-0d25-4ea4-a210-45838580479b/sast/description/131/9017219199198855303)|/src/redis\-cli\.c: [3677](https://github.com/sundb/redis/blob/reload_cgroup_ref//src/redis-cli.c#L3677)|<details><summary>details</summary>The buffer buf created in /src/redis\-cli\.c at line 3677 is written to a buffer in /deps/hiredis/sds\.c at line 234 by newsh, but an error in calc...<br>`ID: %2BpSSxZAM7xfUiads1egmyYebO5I%3D`</details>[Attack Vector](https://gov-il.ast.checkmarx.net/results/3d285292-0d25-4ea4-a210-45838580479b/6e34a6db-7b74-4c41-8bc6-d52b8f2a3534/sast?result-id=%2BpSSxZAM7xfUiads1egmyYebO5I%3D&redirect=true)
![CRITICAL](https://cdn.ast.checkmarx.net/integrations/severity/Critical.png)|[Buffer_Overflow_Wrong_Buffer_Size](https://gov-il.ast.checkmarx.net/results/6e34a6db-7b74-4c41-8bc6-d52b8f2a3534/3d285292-0d25-4ea4-a210-45838580479b/sast/description/131/9017219199198855303)|/src/redis\-cli\.c: [3677](https://github.com/sundb/redis/blob/reload_cgroup_ref//src/redis-cli.c#L3677)|<details><summary>details</summary>The buffer buf created in /src/redis\-cli\.c at line 3677 is written to a buffer in /deps/hiredis/sds\.c at line 234 by hdrlen, but an error in cal...<br>`ID: zN%2FI3F1XTVrKpHuopU6EZZmWXt4%3D`</details>[Attack Vector](https://gov-il.ast.checkmarx.net/results/3d285292-0d25-4ea4-a210-45838580479b/6e34a6db-7b74-4c41-8bc6-d52b8f2a3534/sast?result-id=zN%2FI3F1XTVrKpHuopU6EZZmWXt4%3D&redirect=true)
![CRITICAL](https://cdn.ast.checkmarx.net/integrations/severity/Critical.png)|[Buffer_Overflow_Wrong_Buffer_Size](https://gov-il.ast.checkmarx.net/results/6e34a6db-7b74-4c41-8bc6-d52b8f2a3534/3d285292-0d25-4ea4-a210-45838580479b/sast/description/131/9017219199198855303)|/deps/linenoise/linenoise\.c: [1200](https://github.com/sundb/redis/blob/reload_cgroup_ref//deps/linenoise/linenoise.c#L1200)|<details><summary>details</summary>The buffer buf created in /deps/linenoise/linenoise\.c at line 1200 is written to a buffer in /deps/hiredis/sds\.c at line 97 by sh, but an error i...<br>`ID: oykVSjUcVC%2FEMplDwW4P3YG7%2FzE%3D`</details>[Attack Vector](https://gov-il.ast.checkmarx.net/results/3d285292-0d25-4ea4-a210-45838580479b/6e34a6db-7b74-4c41-8bc6-d52b8f2a3534/sast?result-id=oykVSjUcVC%2FEMplDwW4P3YG7%2FzE%3D&redirect=true)
![CRITICAL](https://cdn.ast.checkmarx.net/integrations/severity/Critical.png)|[Buffer_Overflow_Wrong_Buffer_Size](https://gov-il.ast.checkmarx.net/results/6e34a6db-7b74-4c41-8bc6-d52b8f2a3534/3d285292-0d25-4ea4-a210-45838580479b/sast/description/131/9017219199198855303)|/src/redis\-cli\.c: [10594](https://github.com/sundb/redis/blob/reload_cgroup_ref//src/redis-cli.c#L10594)|<details><summary>details</summary>The buffer argv created in /src/redis\-cli\.c at line 10594 is written to a buffer in /deps/hiredis/sds\.c at line 97 by sh, but an error in calcul...<br>`ID: eStOv%2FTaWfWWBCJCCgzT7mgYJU0%3D`</details>[Attack Vector](https://gov-il.ast.checkmarx.net/results/3d285292-0d25-4ea4-a210-45838580479b/6e34a6db-7b74-4c41-8bc6-d52b8f2a3534/sast?result-id=eStOv%2FTaWfWWBCJCCgzT7mgYJU0%3D&redirect=true)
![CRITICAL](https://cdn.ast.checkmarx.net/integrations/severity/Critical.png)|[Buffer_Overflow_Wrong_Buffer_Size](https://gov-il.ast.checkmarx.net/results/6e34a6db-7b74-4c41-8bc6-d52b8f2a3534/3d285292-0d25-4ea4-a210-45838580479b/sast/description/131/9017219199198855303)|/deps/linenoise/linenoise\.c: [1166](https://github.com/sundb/redis/blob/reload_cgroup_ref//deps/linenoise/linenoise.c#L1166)|<details><summary>details</summary>The buffer fgetc created in /deps/linenoise/linenoise\.c at line 1166 is written to a buffer in /deps/hiredis/sds\.c at line 97 by sh, but an error...<br>`ID: v3h9G7I8PLSutWNyC8k4gGzAdDA%3D`</details>[Attack Vector](https://gov-il.ast.checkmarx.net/results/3d285292-0d25-4ea4-a210-45838580479b/6e34a6db-7b74-4c41-8bc6-d52b8f2a3534/sast?result-id=v3h9G7I8PLSutWNyC8k4gGzAdDA%3D&redirect=true)
![MEDIUM](https://cdn.ast.checkmarx.net/integrations/severity/Medium.png)|[Divide_By_Zero](https://gov-il.ast.checkmarx.net/results/6e34a6db-7b74-4c41-8bc6-d52b8f2a3534/3d285292-0d25-4ea4-a210-45838580479b/sast/description/369/10360811894101193280)|/modules/vector\-sets/fastjson\_test\.c: [121](https://github.com/sundb/redis/blob/reload_cgroup_ref//modules/vector-sets/fastjson_test.c#L121)|<details><summary>details</summary>The application performs an illegal operation in generate\_random\_string, in /modules/vector\-sets/fastjson\_test\.c\. In line 121, the program at...<br>`ID: qiowoZ%2FDUFf8wA3ZCvKY8M0GHks%3D`</details>[Attack Vector](https://gov-il.ast.checkmarx.net/results/3d285292-0d25-4ea4-a210-45838580479b/6e34a6db-7b74-4c41-8bc6-d52b8f2a3534/sast?result-id=qiowoZ%2FDUFf8wA3ZCvKY8M0GHks%3D&redirect=true)
![MEDIUM](https://cdn.ast.checkmarx.net/integrations/severity/Medium.png)|[Divide_By_Zero](https://gov-il.ast.checkmarx.net/results/6e34a6db-7b74-4c41-8bc6-d52b8f2a3534/3d285292-0d25-4ea4-a210-45838580479b/sast/description/369/10360811894101193280)|/src/redis\-cli\.c: [6040](https://github.com/sundb/redis/blob/reload_cgroup_ref//src/redis-cli.c#L6040)|<details><summary>details</summary>The application performs an illegal operation in clusterManagerNodeMasterRandom, in /src/redis\-cli\.c\. In line 6053, the program attempts to divi...<br>`ID: Wdmj3BiFZXbdNClmOY%2Fr1waYywk%3D`</details>[Attack Vector](https://gov-il.ast.checkmarx.net/results/3d285292-0d25-4ea4-a210-45838580479b/6e34a6db-7b74-4c41-8bc6-d52b8f2a3534/sast?result-id=Wdmj3BiFZXbdNClmOY%2Fr1waYywk%3D&redirect=true)

</details>

<!--new issues section--><!--fix issues section--><details><summary><b>Fixed Issues (4)</b></summary>
Great job! The following issues were fixed in this Pull Request

Severity|Issue|Source File / Package
---|---|---
![CRITICAL](https://cdn.ast.checkmarx.net/integrations/severity/Critical.png)|~~[Buffer_Overflow_Wrong_Buffer_Size](https://gov-il.ast.checkmarx.net/results/6e34a6db-7b74-4c41-8bc6-d52b8f2a3534/3d285292-0d25-4ea4-a210-45838580479b/sast/description/131/9017219199198855303)~~|/src/redis\-cli\.c: [3677](https://github.com/sundb/redis/blob/reload_cgroup_ref//src/redis-cli.c#L3677)
![CRITICAL](https://cdn.ast.checkmarx.net/integrations/severity/Critical.png)|~~[Buffer_Overflow_Wrong_Buffer_Size](https://gov-il.ast.checkmarx.net/results/6e34a6db-7b74-4c41-8bc6-d52b8f2a3534/3d285292-0d25-4ea4-a210-45838580479b/sast/description/131/9017219199198855303)~~|/src/redis\-cli\.c: [3677](https://github.com/sundb/redis/blob/reload_cgroup_ref//src/redis-cli.c#L3677)
![CRITICAL](https://cdn.ast.checkmarx.net/integrations/severity/Critical.png)|~~[Buffer_Overflow_Wrong_Buffer_Size](https://gov-il.ast.checkmarx.net/results/6e34a6db-7b74-4c41-8bc6-d52b8f2a3534/3d285292-0d25-4ea4-a210-45838580479b/sast/description/131/9017219199198855303)~~|/src/redis\-cli\.c: [3677](https://github.com/sundb/redis/blob/reload_cgroup_ref//src/redis-cli.c#L3677)
![CRITICAL](https://cdn.ast.checkmarx.net/integrations/severity/Critical.png)|~~[Buffer_Overflow_Wrong_Buffer_Size](https://gov-il.ast.checkmarx.net/results/6e34a6db-7b74-4c41-8bc6-d52b8f2a3534/3d285292-0d25-4ea4-a210-45838580479b/sast/description/131/9017219199198855303)~~|/src/redis\-cli\.c: [3677](https://github.com/sundb/redis/blob/reload_cgroup_ref//src/redis-cli.c#L3677)
</details>

<!--fix issues section-->